refactor: simplification pass — typed settings, ParsedArgs sweep, fit_check split
Removes ~1.5k LOC of partially-migrated patterns.
