+++ src/mkts_backend/cli_tools/asset_check.py
@@ -5,12 +5,11 @@ Resolves a type by ID or partial name, fetches packaged assets for all
 configured characters via ESI, and displays a summary table.
 """
 
-from typing import Dict, List, Optional, Tuple
+from typing import Optional, Tuple
 
 from sqlalchemy import text
 
 from mkts_backend.config.db_config import DatabaseConfig
-from mkts_backend.config.character_config import CharacterConfig, load_characters
 from mkts_backend.config.logging_config import configure_logging
 from mkts_backend.esi.character_assets import fetch_all_character_assets
