+++ src/mkts_backend/cli_tools/fit_check.py
@@ -36,8 +36,6 @@ from mkts_backend.cli_tools.rich_display import (
     print_markdown_export,
     print_overpriced_items,
 )
-from mkts_backend.cli_tools.asset_check import asset_check_command
-
 logger = configure_logging(__name__)
 
 
@@ -1023,582 +1021,26 @@ def display_fit_status(
     return result
 
 
-def _resolve_module_identity(
-    type_id: Optional[int] = None,
-    type_name: Optional[str] = None,
-) -> tuple[int, str]:
-    """
-    Resolve a module to both type_id and type_name via SDE.
-
-    Args:
-        type_id: Type ID to look up (mutually exclusive with type_name)
-        type_name: Type name to search for (mutually exclusive with type_id)
-
-    Returns:
-        Tuple of (type_id, type_name)
-
-    Raises:
-        ValueError: If the module cannot be resolved or is ambiguous
-    """
-    sde_db = DatabaseConfig("sde")
-
-    if type_id is not None:
-        with sde_db.engine.connect() as conn:
-            query = text(
-                "SELECT typeName FROM sdetypes WHERE typeID = :type_id")
-            result = conn.execute(query, {"type_id": type_id}).fetchone()
-            if result:
-                return type_id, result[0]
-            raise ValueError(f"No item found with typeID={type_id}")
-
-    if type_name is not None:
-        with sde_db.engine.connect() as conn:
-            # Try exact match first
-            query = text(
-                "SELECT typeID, typeName FROM sdetypes WHERE typeName = :name"
-            )
-            result = conn.execute(query, {"name": type_name}).fetchone()
-            if result:
-                return result[0], result[1]
-
-            # Try partial match
-            query = text(
-                "SELECT typeID, typeName FROM sdetypes "
-                "WHERE typeName LIKE :pattern ORDER BY typeName LIMIT 10"
-            )
-            rows = conn.execute(
-                query, {"pattern": f"%{type_name}%"}).fetchall()
-            if len(rows) == 1:
-                return rows[0][0], rows[0][1]
-            elif len(rows) > 1:
-                names = "\n  ".join(
-                    f"{r[0]}: {r[1]}" for r in rows)
-                raise ValueError(
-                    f"Ambiguous name '{type_name}'. Matches:\n  {names}"
-                )
-            raise ValueError(f"No item found matching '{type_name}'")
-
-    raise ValueError("Either --id or --name is required")
-
-
-def _query_module_usage(
-    type_id: int,
-    market_ctx: Optional[MarketContext] = None,
-) -> List[Dict]:
-    """
-    Query which fits use a given module and their market status.
-
-    Joins doctrines and doctrine_fits on fit_id where type_id matches.
-
-    Args:
-        type_id: The type ID of the module to look up
-        market_ctx: Market context for database selection
-
-    Returns:
-        List of dicts with fit usage and market data
-    """
-    db_alias = market_ctx.database_alias if market_ctx else "wcmkt"
-    db = DatabaseConfig(db_alias)
-
-    results = []
-    with db.engine.connect() as conn:
-        query = text("""
-            SELECT
-                d.fit_id,
-                df.fit_name,
-                df.ship_name,
-                df.doctrine_name,
-                d.fit_qty,
-                df.target,
-                d.total_stock,
-                d.fits_on_mkt,
-                d.price
-            FROM doctrines d
-            JOIN doctrine_fits df ON d.fit_id = df.fit_id
-            WHERE d.type_id = :type_id
-            ORDER BY df.doctrine_name, df.fit_name
-        """)
-        rows = conn.execute(query, {"type_id": type_id}).fetchall()
-
-        for row in rows:
-            fits = row.fits_on_mkt or 0
-            target = row.target or 0
-            fit_qty = row.fit_qty or 1
-            qty_needed = max(0, int(
-                (target - fits) * fit_qty)) if fits < target else 0
-
-            results.append({
-                "fit_id": row.fit_id,
-                "fit_name": row.fit_name,
-                "ship_name": row.ship_name,
-                "doctrine_name": row.doctrine_name,
-                "fit_qty": fit_qty,
-                "target": target,
-                "total_stock": row.total_stock or 0,
-                "fits_on_mkt": fits,
-                "qty_needed": qty_needed,
-                "price": row.price,
-            })
-
-    return results
-
-
-def module_command(
-    type_id: Optional[int] = None,
-    type_name: Optional[str] = None,
-    market_alias: str = "primary",
-) -> bool:
-    """
-    Display which fits use a module and their market availability.
-
-    Args:
-        type_id: Type ID of the module
-        type_name: Type name of the module (alternative to type_id)
-        market_alias: Market alias or "both" for dual-market display
-
-    Returns:
-        True if successful, False otherwise
-    """
-    # Resolve module identity
-    try:
-        resolved_id, resolved_name = _resolve_module_identity(
-            type_id, type_name)
-    except ValueError as e:
-        console.print(f"[red]Error: {e}[/red]")
-        return False
-
-    show_both = market_alias == "both"
-
-    if show_both:
-        # Query both markets
-        try:
-            primary_ctx = MarketContext.from_settings("primary")
-            deploy_ctx = MarketContext.from_settings("deployment")
-        except ValueError as e:
-            console.print(f"[red]Error: {e}[/red]")
-            return False
-
-        primary_data = _query_module_usage(resolved_id, primary_ctx)
-        deploy_data = _query_module_usage(resolved_id, deploy_ctx)
-
-        if not primary_data and not deploy_data:
-            console.print(
-                f"[yellow]Module '{resolved_name}' (ID: {resolved_id}) "
-                f"is not used in any tracked fits.[/yellow]"
-            )
-            return True
-
-        # Merge data: key by fit_id
-        deploy_by_fit = {r["fit_id"]: r for r in deploy_data}
-        merged = []
-        seen_fit_ids = set()
-
-        for row in primary_data:
-            fid = row["fit_id"]
-            seen_fit_ids.add(fid)
-            d_row = deploy_by_fit.get(fid, {})
-            merged.append({
-                "fit_id": fid,
-                "fit_name": row["fit_name"],
-                "ship_name": row["ship_name"],
-                "doctrine_name": row["doctrine_name"],
-                "fit_qty": row["fit_qty"],
-                "target": row["target"],
-                "price": row["price"],
-                "p_stock": row["total_stock"],
-                "p_fits": row["fits_on_mkt"],
-                "p_need": row["qty_needed"],
-                "d_stock": d_row.get("total_stock", 0),
-                "d_fits": d_row.get("fits_on_mkt", 0),
-                "d_need": d_row.get("qty_needed", 0),
-            })
-
-        # Add any deployment-only fits
-        for fid, d_row in deploy_by_fit.items():
-            if fid not in seen_fit_ids:
-                merged.append({
-                    "fit_id": fid,
-                    "fit_name": d_row["fit_name"],
-                    "ship_name": d_row["ship_name"],
-                    "doctrine_name": d_row["doctrine_name"],
-                    "fit_qty": d_row["fit_qty"],
-                    "target": d_row["target"],
-                    "price": d_row["price"],
-                    "p_stock": 0,
-                    "p_fits": 0,
-                    "p_need": 0,
-                    "d_stock": d_row["total_stock"],
-                    "d_fits": d_row["fits_on_mkt"],
-                    "d_need": d_row["qty_needed"],
-                })
-
-        table = create_module_usage_table(
-            resolved_name, resolved_id, merged, show_both=True)
-        console.print(table)
-        console.print(
-            f"\n[dim]Total: {len(merged)} fit(s) using this module[/dim]")
-
-    else:
-        # Single market
-        try:
-            market_ctx = MarketContext.from_settings(market_alias)
-        except ValueError as e:
-            console.print(f"[red]Error: {e}[/red]")
-            console.print(f"Available markets: {
-                ', '.join(MarketContext.list_available())}")
-            return False
-
-        data = _query_module_usage(resolved_id, market_ctx)
-
-        if not data:
-            console.print(
-                f"[yellow]Module '{resolved_name}' (ID: {resolved_id}) "
-                f"is not used in any tracked fits on {market_ctx.name}.[/yellow]"
-            )
-            return True
-
-        table = create_module_usage_table(
-            resolved_name, resolved_id, data, show_both=False)
-        console.print(table)
-        console.print(
-            f"\n[dim]Total: {len(data)} fit(s) using this module "
-            f"on {market_ctx.name}[/dim]"
-        )
-
-    return True
-
-
-def _query_needed_data(
-    market_ctx: Optional[MarketContext] = None,
-    ship_filter: Optional[List[str]] = None,
-    fit_filter: Optional[List[int]] = None,
-    targ_perc_filter: Optional[float] = None,
-) -> List[Dict]:
-    """
-    Query needed items by joining doctrines and ship_targets tables.
-
-    Args:
-        market_ctx: Market context for database selection
-        ship_filter: Filter by ship names (case-insensitive substring match, any must match)
-        fit_filter: Filter by fit_ids (any must match)
-        targ_perc_filter: Filter where targ_perc < this value
-
-    Returns:
-        List of dicts with needed item data
-    """
-    db_alias = market_ctx.database_alias if market_ctx else "wcmkt"
-    db = DatabaseConfig(db_alias)
-
-    results = []
-    with db.engine.connect() as conn:
-        query = text("""
-            SELECT
-                d.fit_id,
-                d.ship_id,
-                d.ship_name,
-                d.type_id,
-                d.type_name,
-                d.fit_qty,
-                t.ship_target AS target,
-                t.fit_name,
-                d.fits_on_mkt,
-                d.total_stock,
-                round((1.0 * d.fits_on_mkt) / NULLIF(t.ship_target, 0), 2) AS targ_perc,
-                CASE
-                    WHEN d.fits_on_mkt < t.ship_target
-                        THEN (NULLIF(t.ship_target, 0) - d.fits_on_mkt) * d.fit_qty
-                    ELSE 0
-                END AS qty_needed
-            FROM doctrines AS d
-            LEFT JOIN ship_targets AS t
-                ON d.fit_id = t.fit_id
-            WHERE CASE
-                    WHEN d.fits_on_mkt < t.ship_target
-                        THEN (NULLIF(t.ship_target, 0) - d.fits_on_mkt) * d.fit_qty
-                    ELSE 0
-                  END > 0
-            ORDER BY d.ship_name, d.fit_id, targ_perc
-        """)
-        rows = conn.execute(query).fetchall()
-
-        for row in rows:
-            item = {
-                "fit_id": row.fit_id,
-                "ship_id": row.ship_id,
-                "ship_name": row.ship_name,
-                "type_id": row.type_id,
-                "type_name": row.type_name,
-                "fit_qty": row.fit_qty or 1,
-                "target": row.target,
-                "fit_name": row.fit_name or "Unknown",
-                "fits_on_mkt": row.fits_on_mkt or 0,
-                "total_stock": row.total_stock or 0,
-                "targ_perc": row.targ_perc or 0,
-                "qty_needed": row.qty_needed or 0,
-            }
-
-            # Apply filters (exact match — names are pre-resolved by _resolve_ship_filters)
-            if ship_filter and item["ship_name"] not in ship_filter:
-                continue
-            if fit_filter and item["fit_id"] not in fit_filter:
-                continue
-
-            results.append(item)
-
-    # Apply equivalent module stock
-    all_type_ids = list({item["type_id"] for item in results})
-    equiv_stock = get_equiv_stock(all_type_ids, market_ctx)
-
-    for item in results:
-        equivs = equiv_stock.get(item["type_id"])
-        if not equivs:
-            continue
-
-        equiv_total = sum(e["stock"] for e in equivs)
-        item["total_stock"] = (item.get("total_stock", 0) or 0) + equiv_total
-        target_val = item.get("target", 0) or 0
-        fit_qty = item.get("fit_qty", 1) or 1
-
-        # Recalculate fits_on_mkt from updated total_stock
-        # doctrines stores fits_on_mkt = total_stock / fit_qty
-        item["fits_on_mkt"] = item["total_stock"] / fit_qty if fit_qty > 0 else 0
-        item["targ_perc"] = round(item["fits_on_mkt"] / target_val, 2) if target_val > 0 else 0
-        item["qty_needed"] = max(0, int((target_val - item["fits_on_mkt"]) * fit_qty)) if item["fits_on_mkt"] < target_val else 0
-        item["equiv_items"] = equivs
-
-    # Apply targ_perc filter after equiv adjustment
-    if targ_perc_filter is not None:
-        results = [item for item in results if item["targ_perc"] < targ_perc_filter]
-
-    # Remove items that no longer need anything after equiv adjustment
-    results = [item for item in results if item["qty_needed"] > 0]
-
-    return results
-
-
-def needed_command(
-    market_alias: str = "primary",
-    ship_filter: Optional[List[str]] = None,
-    fit_filter: Optional[List[int]] = None,
-    targ_perc_filter: Optional[float] = None,
-    show_assets: bool = False,
-    force_refresh: bool = False,
-) -> bool:
-    """
-    Display needed items organized into sub-tables by fit.
-
-    Args:
-        market_alias: Market alias (primary, deployment)
-        ship_filter: Filter by ship names (comma-separated or repeated flags)
-        fit_filter: Filter by fit_ids (comma-separated or repeated flags)
-        targ_perc_filter: Filter where targ_perc < this value
-        show_assets: If True, fetch and display per-character asset columns
-
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        market_ctx = MarketContext.from_settings(market_alias)
-    except ValueError as e:
-        console.print(f"[red]Error: {e}[/red]")
-        return False
-
-    data = _query_needed_data(
-        market_ctx,
-        ship_filter=ship_filter,
-        fit_filter=fit_filter,
-        targ_perc_filter=targ_perc_filter,
-    )
-
-    if not data:
-        console.print("[yellow]No items needed - all fits are at or above target.[/yellow]")
-        return True
-
-    # Fetch per-character assets if requested
-    char_assets = None
-    if show_assets:
-        from mkts_backend.esi.character_assets import fetch_all_character_assets
-
-        all_type_ids = list({item["type_id"] for item in data})
-        char_assets = fetch_all_character_assets(
-            type_ids=all_type_ids, force_refresh=force_refresh
-        )
-
-    # Group by fit_id
-    grouped = {}
-    for item in data:
-        fid = item["fit_id"]
-        if fid not in grouped:
-            grouped[fid] = []
-        grouped[fid].append(item)
-
-    # Print header
-    console.print()
-    filter_parts = []
-    if ship_filter:
-        filter_parts.append(f"ship={','.join(ship_filter)}")
-    if fit_filter:
-        filter_parts.append(f"fit={','.join(str(f) for f in fit_filter)}")
-    if targ_perc_filter is not None:
-        filter_parts.append(f"targ_perc<{targ_perc_filter}")
-    filter_str = f" ({', '.join(filter_parts)})" if filter_parts else ""
-
-    console.print(
-        f"[bold]Needed Items[/bold] - [green]{market_ctx.name}[/green]{filter_str}",
-    )
-    console.print()
-
-    # Render a sub-table for each fit group
-    for fit_id, items in grouped.items():
-        ship_id = items[0].get("ship_id")
-        ship_name = items[0]["ship_name"]
-        fit_name = items[0]["fit_name"]
-        target = items[0]["target"] or 0
-
-        # min(fits_on_mkt) across all items in this fit
-        min_fits = min((item["fits_on_mkt"] for item in items), default=0)
-
-        table = create_needed_table(
-            fit_id=fit_id,
-            ship_id=ship_id,
-            ship_name=ship_name,
-            fit_name=fit_name,
-            min_fits=min_fits,
-            target=target,
-            items=items,
-            char_assets=char_assets,
-        )
-        console.print(table)
-        console.print()
-
-    # Summary
-    total_fits = len(grouped)
-    total_items = len(data)
-    console.print(
-        f"[dim]{total_fits} fit(s), {total_items} item(s) below target[/dim]"
-    )
-
-    return True
-
-def __handle_assets(type_id: int, type_name: str, force_refresh = False):
-    asset_check_command(type_id, type_name, force_refresh)
-    
-
-def _resolve_ship_filters(
-    ship_inputs: List[str],
-    market_ctx: Optional[MarketContext] = None,
-) -> Optional[List[str]]:
-    """
-    Resolve user ship name inputs to exact ship names via fuzzy matching.
-
-    For each input, finds substring matches among distinct ship names in
-    doctrine_fits. If one match, auto-selects. If multiple, prompts the
-    user to choose.
-
-    Args:
-        ship_inputs: Raw ship name fragments from CLI (e.g. ["Hurricane"])
-        market_ctx: Market context for database selection
-
-    Returns:
-        List of resolved exact ship names, or None if user cancels
-    """
-    from rich.prompt import Prompt
-
-    db_alias = market_ctx.database_alias if market_ctx else "wcmkt"
-    db = DatabaseConfig(db_alias)
-
-    with db.engine.connect() as conn:
-        query = text("SELECT DISTINCT ship_name FROM doctrine_fits ORDER BY ship_name")
-        all_ships = [row[0] for row in conn.execute(query).fetchall()]
-
-    resolved: List[str] = []
-
-    for fragment in ship_inputs:
-        matches = [s for s in all_ships if fragment.lower() in s.lower()]
-
-        if not matches:
-            console.print(f"[yellow]No ships matching '{fragment}'[/yellow]")
-            continue
-        elif len(matches) == 1:
-            resolved.append(matches[0])
-        else:
-            # Exact match takes priority (e.g. "Hurricane" matches "Hurricane" exactly)
-            exact = [s for s in matches if s.lower() == fragment.lower()]
-            if exact:
-                # Still show the options since the user might want a variant
-                pass
-
-            console.print(f"\n[bold]Ships matching '{fragment}':[/bold]")
-            console.print("  [dim]0[/dim]) All matches")
-            for i, name in enumerate(matches, 1):
-                console.print(f"  [dim]{i}[/dim]) {name}")
-
-            choice = Prompt.ask(
-                "Select",
-                default="0",
-            )
-
-            try:
-                idx = int(choice)
-            except ValueError:
-                console.print("[red]Invalid selection, skipping[/red]")
-                continue
-
-            if idx == 0:
-                resolved.extend(matches)
-            elif 1 <= idx <= len(matches):
-                resolved.append(matches[idx - 1])
-            else:
-                console.print("[red]Invalid selection, skipping[/red]")
-                continue
-
-    return resolved if resolved else None
-
-
-def _handle_needed(sub_args: List[str]) -> None:
-    """
-    Handle the needed subcommand.
-
-    Args:
-        sub_args: Remaining arguments after 'needed'
-    """
-    from mkts_backend.cli_tools.arg_utils import ParsedArgs, ArgError
-
-    p = ParsedArgs(sub_args)
-    market_alias = parse_market_args(sub_args)
-    show_assets = p.has_flag("assets")
-    force_refresh = p.has_flag("refresh")
-
-    ship_filters = p.get_string_list("ship")
-
-    try:
-        fit_filters = p.get_int_list("fit", "fit-id", "fit_id", "id")
-        targ_perc_filter = p.get_float("target")
-    except ArgError as e:
-        console.print(f"[red]Error: {e}[/red]")
-        return
-
-    # Resolve fuzzy ship names to exact matches
-    resolved_ships = None
-    if ship_filters:
-        try:
-            market_ctx = MarketContext.from_settings(market_alias)
-        except ValueError as e:
-            console.print(f"[red]Error: {e}[/red]")
-            return
-        resolved_ships = _resolve_ship_filters(ship_filters, market_ctx)
-        if resolved_ships is None:
-            console.print("[yellow]No matching ships found.[/yellow]")
-            return
-
-    needed_command(
-        market_alias=market_alias,
-        ship_filter=resolved_ships,
-        fit_filter=fit_filters or None,
-        targ_perc_filter=targ_perc_filter,
-        show_assets=show_assets,
-        force_refresh=force_refresh,
-    )
+# ── module subcommand ──────────────────────────────────────────
+# Implementation lives in fit_check_module.py; re-exported here so existing
+# imports (e.g. `from mkts_backend.cli_tools.fit_check import module_command`)
+# keep working.
+from mkts_backend.cli_tools.fit_check_module import (  # noqa: E402
+    _resolve_module_identity,
+    _query_module_usage,
+    module_command,
+)
+
+
+# ── needed subcommand ──────────────────────────────────────────
+# Implementation lives in fit_check_needed.py; re-exported for backward
+# compatibility with existing imports.
+from mkts_backend.cli_tools.fit_check_needed import (  # noqa: E402, F401
+    _query_needed_data,
+    _resolve_ship_filters,
+    needed_command,
+)
+from mkts_backend.cli_tools.fit_check_needed import handle_needed as _handle_needed  # noqa: E402, F401
 
 
 def _handle_list_fits(sub_args: List[str]) -> None:
@@ -1622,35 +1064,7 @@ def _handle_list_fits(sub_args: List[str]) -> None:
     list_fits_command(db_alias=db_alias)
 
 
-def _handle_module(sub_args: List[str]) -> None:
-    """
-    Handle the module subcommand.
-
-    Args:
-        sub_args: Remaining arguments after 'module'
-    """
-    from mkts_backend.cli_tools.arg_utils import ParsedArgs, ArgError
-
-    p = ParsedArgs(sub_args)
-    market_alias = parse_market_args(sub_args)
-
-    try:
-        type_id = p.get_int("id")
-    except ArgError as e:
-        console.print(f"[red]Error: {e}[/red]")
-        return
-
-    type_name = p.get_string("name")
-
-    if type_id is None and type_name is None:
-        console.print(
-            "[red]Error: --id=<type_id> or --name=<name> is required[/red]")
-        console.print("Usage: fitcheck module --id=11269")
-        console.print('       fitcheck module --name="Multispectrum Energized Membrane II"')
-        return
-
-    module_command(type_id=type_id, type_name=type_name,
-                   market_alias=market_alias)
+from mkts_backend.cli_tools.fit_check_module import handle_module as _handle_module  # noqa: E402, F401
 
 
 def _handle_assets(sub_args: List[str]) -> None:
